requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
easygui==0.98.1
future==0.18.2
gast==0.3.3
grpcio==1.27.2
h5py==2.10.0
iso8601==0.1.12
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pygame==1.9.6
pymongo==3.10.1
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyserial==3.4
python-dateutil==2.8.1
PyYAML==5.3.1
sklearn==0.0
scipy==1.4.1
six==1.14.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==1.0.0
